refactor: rm legacy params from all non-sdk funcs
