+++ modules/apps/27-interchain-accounts/controller/types/params.go
@@ -5,6 +5,8 @@ const (
 	DefaultControllerEnabled = true
 )
 
+var KeyControllerEnabled = []byte("ControllerEnabled")
+
 // NewParams creates a new parameter configuration for the controller submodule
 func NewParams(enableController bool) Params {
 	return Params{
