--- modules/apps/27-interchain-accounts/host/types/params_legacy.go
@@ -1,59 +1,0 @@
-/*
-NOTE: Usage of x/params to manage parameters is deprecated in favor of x/gov
-controlled execution of MsgUpdateParams messages. These types remains solely
-for migration purposes and will be removed in a future release.
-[#3621](https://github.com/cosmos/ibc-go/issues/3621)
-*/
-
-package types
-
-import (
-	"fmt"
-	"strings"
-
-	paramtypes "github.com/cosmos/cosmos-sdk/x/params/types"
-)
-
-var (
-	// KeyHostEnabled is the store key for HostEnabled Params
-	KeyHostEnabled = []byte("HostEnabled")
-	// KeyAllowMessages is the store key for the AllowMessages Params
-	KeyAllowMessages = []byte("AllowMessages")
-)
-
-// ParamKeyTable type declaration for parameters
-func ParamKeyTable() paramtypes.KeyTable {
-	return paramtypes.NewKeyTable().RegisterParamSet(&Params{})
-}
-
-// ParamSetPairs implements params.ParamSet
-func (p *Params) ParamSetPairs() paramtypes.ParamSetPairs {
-	return paramtypes.ParamSetPairs{
-		paramtypes.NewParamSetPair(KeyHostEnabled, &p.HostEnabled, validateEnabledType),
-		paramtypes.NewParamSetPair(KeyAllowMessages, &p.AllowMessages, validateAllowlistLegacy),
-	}
-}
-
-func validateEnabledType(i any) error {
-	_, ok := i.(bool)
-	if !ok {
-		return fmt.Errorf("invalid parameter type: %T", i)
-	}
-
-	return nil
-}
-
-func validateAllowlistLegacy(i any) error {
-	allowMsgs, ok := i.([]string)
-	if !ok {
-		return fmt.Errorf("invalid parameter type: %T", i)
-	}
-
-	for _, typeURL := range allowMsgs {
-		if strings.TrimSpace(typeURL) == "" {
-			return fmt.Errorf("parameter must not contain empty strings: %s", allowMsgs)
-		}
-	}
-
-	return nil
-}
