+++ modules/apps/transfer/types/params.go
@@ -7,6 +7,13 @@ const (
 	DefaultReceiveEnabled = true
 )
 
+var (
+	// KeySendEnabled is store's key for SendEnabled Params
+	KeySendEnabled = []byte("SendEnabled")
+	// KeyReceiveEnabled is store's key for ReceiveEnabled Params
+	KeyReceiveEnabled = []byte("ReceiveEnabled")
+)
+
 // NewParams creates a new parameter configuration for the ibc transfer module
 func NewParams(enableSend, enableReceive bool) Params {
 	return Params{
