--- modules/apps/transfer/types/params_legacy.go
@@ -1,42 +1,0 @@
-/*
-NOTE: Usage of x/params to manage parameters is deprecated in favor of x/gov
-controlled execution of MsgUpdateParams messages. These types remains solely
-for migration purposes and will be removed in a future release.
-[#3621](https://github.com/cosmos/ibc-go/issues/3621)
-*/
-package types
-
-import (
-	"fmt"
-
-	paramtypes "github.com/cosmos/cosmos-sdk/x/params/types"
-)
-
-var (
-	// KeySendEnabled is store's key for SendEnabled Params
-	KeySendEnabled = []byte("SendEnabled")
-	// KeyReceiveEnabled is store's key for ReceiveEnabled Params
-	KeyReceiveEnabled = []byte("ReceiveEnabled")
-)
-
-// ParamKeyTable type declaration for parameters
-func ParamKeyTable() paramtypes.KeyTable {
-	return paramtypes.NewKeyTable().RegisterParamSet(&Params{})
-}
-
-// ParamSetPairs implements params.ParamSet
-func (p *Params) ParamSetPairs() paramtypes.ParamSetPairs {
-	return paramtypes.ParamSetPairs{
-		paramtypes.NewParamSetPair(KeySendEnabled, &p.SendEnabled, validateEnabledTypeLegacy),
-		paramtypes.NewParamSetPair(KeyReceiveEnabled, &p.ReceiveEnabled, validateEnabledTypeLegacy),
-	}
-}
-
-func validateEnabledTypeLegacy(i any) error {
-	_, ok := i.(bool)
-	if !ok {
-		return fmt.Errorf("invalid parameter type: %T", i)
-	}
-
-	return nil
-}
