--- modules/core/02-client/types/params_legacy.go
@@ -1,37 +1,0 @@
-/*
-NOTE: Usage of x/params to manage parameters is deprecated in favor of x/gov
-controlled execution of MsgUpdateParams messages. These types remains solely
-for migration purposes and will be removed in a future release.
-[#3621](https://github.com/cosmos/ibc-go/issues/3621)
-*/
-package types
-
-import (
-	"fmt"
-
-	paramtypes "github.com/cosmos/cosmos-sdk/x/params/types"
-)
-
-// KeyAllowedClients is store's key for AllowedClients Params
-var KeyAllowedClients = []byte("AllowedClients")
-
-// ParamKeyTable type declaration for parameters
-func ParamKeyTable() paramtypes.KeyTable {
-	return paramtypes.NewKeyTable().RegisterParamSet(&Params{})
-}
-
-// ParamSetPairs implements params.ParamSet
-func (p *Params) ParamSetPairs() paramtypes.ParamSetPairs {
-	return paramtypes.ParamSetPairs{
-		paramtypes.NewParamSetPair(KeyAllowedClients, &p.AllowedClients, validateClientsLegacy),
-	}
-}
-
-func validateClientsLegacy(i any) error {
-	clients, ok := i.([]string)
-	if !ok {
-		return fmt.Errorf("invalid parameter type: %T", i)
-	}
-
-	return validateClients(clients)
-}
